WIP: refactoring object details panel
