+++ src/components/scratchpad.jsx
@@ -1,10 +1,12 @@
 import { useRef } from 'react';
 import { Link, useNavigate, Outlet } from 'react-router-dom';
+import MetaDataTable from '../components/tables/MetaDataTable';
 import UseEscape from '../hooks/useEscape';
+import { HiOutlinePencilSquare, HiOutlineArrowDownTray, HiOutlineTrash, HiXMark } from 'react-icons/hi2';
 import useOnClickOutside from '../hooks/useOnClickOutside';
-import ObjectPreview from '../components/ObjectPreview';
+import ObjectPreview from './ObjectPreview';
 
-export default function ObjectDetials({ accountId, authKey }) {
+export default function DisplayObject({ accountId, authKey, objectData, container, object, setMode }) {
   const [mode, setMode] = useState('display');
 
   let { container, object } = useParams();
@@ -65,42 +67,4 @@ export default function ObjectDetials({ accountId, authKey }) {
       </div>
     </>
   );
-}
-
-// Called to load data for any GET request
-export async function loader(params, accountId, x_auth_token) {
-  const response = await fetch(
-    `https://api.testnet.onmachina.io/v1/${accountId}/${params.container}/${params.object}`,
-    {
-      method: 'HEAD',
-      headers: {
-        'x-auth-token': x_auth_token,
-      },
-    },
-  );
-  const headersArray = [];
-  for (const [name, value] of response.headers.entries()) {
-    if (name === 'last-modified') headersArray.push({ name, value });
-    if (name === 'content-type') headersArray.push({ name, value });
-    if (name === 'etag') headersArray.push({ name, value });
-    if (name.startsWith('x-object-meta')) headersArray.push({ name, value });
-  }
-  return headersArray;
-}
-
-// Called for any POST request
-export async function action({ request, params }) {
-  const formData = await request.formData();
-  const action = Object.fromEntries(formData).action;
-  const token = Object.fromEntries(formData).token;
-  const accountId = Object.fromEntries(formData).accountId;
-  if (action === 'deleteObject') {
-    await deleteObject(params.container, params.object, accountId, token);
-    return redirect(`/${params.container}`);
-  } else if (action === 'rename') {
-    const oldName = Object.fromEntries(formData).oldname;
-    const newName = Object.fromEntries(formData).newname;
-    await renameObject({ container: params.container, oldName, newName, accountId, token });
-    return redirect(`/${params.container}/${newName}`);
-  }
 }
